Move methods to new ServiceBase class
